[Editorial] Update the "Payload Formats" section
Incorporate use of the new terms "data item chunk" and "body chunk".

Move the "Payload Formats" subsection into the "Transport Messages"
chapter, since the information in "Payload Formats" depends on the
definitions in "Transport Messages". This also moves the "Payload
Formats" and the "Choosing a Reply Payload Format" sections
together.