[Technical] Replace iWARP terminology
During AUTH48 for RFC 8797, Tom Talpey observed that the use of the
term "iWARP" was deprecated or inappropriate. The current document
uses that term the same way, thus needs a similar update.

Also, "Reliable Connected" is an InfiniBand term. The text is made
more agnostic, as it was done for RFC 8797, now that the document
includes a citable reference to RFC 5044.