[Technical] Replace the RDMA2_MSG and RDMA2_NOMSG header types
We've observed that the function and content of the chunk lists is
different between Call messages and Reply messages. Rather than
setting a flag in the header, split the MSG header type into
one for Calls and one for Replies. Likewise for NOMSG headers.

Make a similar adjustment for continued messages and the CONNPROP
headers.

Now that we have a separate CALL_EXTERNAL header type, replace the
double use of the Read list for both data item chunks and body
chunks by adding a separate list just for the Position zero body
chunk.

As a result of these changes, we can simplify the data structures
that define the Call and Reply header type "arguments", and remove
the rdma_flags field.