[Technical] Add a second VERS-related error code
RDMA2_ERR_VERS indicates when a Receiver does not recognize the
transport protocol version in an incoming message. We need another
error code to indicate when a Receiver /does/ recognize the
protocol version, but it does not match the version fixed by
previous messages on the connection.

For example, this error code can be used if a reverse-direction
message uses a different protocol version than previous forward-
direction messages.

In a v1-only environment, ERR_VERS sufficed for both situations
because the only allowed protocol version is 1. There is no
distinction between the two situations.