[Technical] Add language covering Call-only operation
Call-only means there's no co-responding Reply.

- The document already covers the issues around credit management
for this style of operation

- But note that Call-only cannot use any explicit RDMA resources,
since there's no indication from the Responder that it has finished
using Read chunks and the Requester may invalidate them.

We could re-introduce RDMA_DONE for this purpose, but generally
speaking Call-only procedures are quite rare, and typically small
enough to be handled via a single or a few RDMA Sends.